+++ crates/biome_js_analyze/src/lint/nursery/no_useless_undefined_initialization.rs
@@ -3,7 +3,9 @@ use biome_analyze::{
     RuleSource, RuleSourceKind,
 };
 use biome_console::markup;
-use biome_js_syntax::JsVariableDeclaration;
+use biome_js_factory::make::js_variable_declarator_list;
+use biome_js_syntax::{JsLanguage, JsVariableDeclarator, JsVariableStatement};
+use biome_rowan::{chain_trivia_pieces, SyntaxToken, SyntaxTriviaPiece};
 use biome_rowan::{AstNode, BatchMutationExt, TextRange};
 
 use crate::JsRuleAction;
@@ -13,8 +15,8 @@ declare_rule! {
     ///
     /// A variable that is declared and not initialized to any value automatically gets the value of `undefined`.
     /// It’s considered a best practice to avoid initializing variables to `undefined`.
-    /// Please note that any inline comments attached to the initialization value or variable will be removed on auto-fix.
-    /// Please be also aware that this differs from Eslint's behaviour and we are still discussing on how to properly handle this case.
+    /// Please note that any inline comments attached to the initialization value or variable will be moved at the end of the variable declaration on auto-fix.
+    /// Please be also aware that this differs from Eslint's behaviour.
     ///
     /// ## Examples
     ///
@@ -51,22 +53,26 @@ declare_rule! {
         language: "js",
         sources: &[RuleSource::Eslint("no-undef-init")],
         source_kind: RuleSourceKind::Inspired,
-        fix_kind: FixKind::Unsafe,
+        fix_kind: FixKind::Safe,
         recommended: false,
     }
 }
 
 impl Rule for NoUselessUndefinedInitialization {
-    type Query = Ast<JsVariableDeclaration>;
+    type Query = Ast<JsVariableStatement>;
     type State = (String, TextRange);
     type Signals = Vec<Self::State>;
     type Options = ();
 
     fn run(ctx: &RuleContext<Self>) -> Self::Signals {
-        let node = ctx.query();
-        let let_or_var_kind = node.is_let() || node.is_var();
+        let statement = ctx.query();
 
         let mut signals = vec![];
+        let Ok(node) = statement.declaration() else {
+            return signals;
+        };
+
+        let let_or_var_kind = node.is_let() || node.is_var();
 
         if !let_or_var_kind {
             return signals;
@@ -110,23 +116,70 @@ impl Rule for NoUselessUndefinedInitialization {
     }
 
     fn action(ctx: &RuleContext<Self>, state: &Self::State) -> Option<JsRuleAction> {
-        let declarators = ctx.query().declarators();
+        let node = ctx.query();
+        let assignment_statement = node.clone();
 
-        let initializer = declarators
+        let current_declaration_statement = node.clone().declaration().ok()?;
+        let declarators = current_declaration_statement.declarators();
+
+        let current_declaration = declarators
             .clone()
             .into_iter()
-            .find(|el| {
-                el.as_ref()
-                    .ok()
-                    .and_then(|element| element.id().ok())
-                    .is_some_and(|id| id.text() == state.0)
+            .flat_map(|declarator| declarator.ok())
+            .find(|decl| decl.id().is_ok_and(|id| id.text() == state.0))?;
+
+        let current_initializer = current_declaration.initializer()?;
+
+        let eq_token_trivia = current_initializer
+            .eq_token()
+            .map(|token| token.trailing_trivia())
+            .ok()?
+            .pieces();
+
+        let expression_trivia = current_initializer
+            .expression()
+            .ok()?
+            .as_js_reference_identifier()
+            .map(|reference| reference.value_token())?
+            .ok()?
+            .trailing_trivia()
+            .pieces();
+
+        // Save the separators too
+        let separators_syntax = declarators.clone().into_syntax();
+        let separators: Vec<SyntaxToken<JsLanguage>> = separators_syntax.tokens().collect();
+
+        let new_declaration = current_declaration.clone().with_initializer(None);
+        let new_declarators: Vec<JsVariableDeclarator> = declarators
+            .clone()
+            .into_iter()
+            .flat_map(|decl| decl.ok())
+            .map(|decl| {
+                if decl == current_declaration {
+                    new_declaration.clone()
+                } else {
+                    decl
+                }
             })
-            .map(|decl| decl.ok())?
-            .and_then(|declarator| declarator.initializer())?;
+            .collect();
+
+        // Recreate the declaration statement with updated declarators
+        let new_declaration_statement = current_declaration_statement
+            .with_declarators(js_variable_declarator_list(new_declarators, separators));
+
+        let chained_comments: Vec<SyntaxTriviaPiece<JsLanguage>> =
+            chain_trivia_pieces(eq_token_trivia, expression_trivia)
+                .filter(|trivia| trivia.is_comments())
+                .collect();
+
+        // Create the whole statement using updated subtree and append comments to the statement
+        let new_node = assignment_statement
+            .clone()
+            .with_declaration(new_declaration_statement)
+            .append_trivia_pieces(chained_comments)?;
 
         let mut mutation = ctx.root().begin();
-        // This will remove any comments attached to the initialization clause
-        mutation.remove_node(initializer);
+        mutation.replace_node_discard_trivia(assignment_statement, new_node);
 
         Some(JsRuleAction::new(
             ActionCategory::QuickFix,